core_service/requirements.txt
django>=4.2
psycopg2-binary
python-dotenv
Pillow
djangorestframework
djangorestframework-simplejwt
